Make this work with Ultimaker API v8; logging
Remove version specification on QtQuick.Controls: v1.1 doesn't seem
to be in the Cura 8.0 API, so specifying it breaks the plugin.
Removing the version specification should make it load the latest.
Possibly should do this for the other imports.

Roll the PyQt versions to PyQt6 and change the reference to the Key
enum as necessary.

Added logging and left it in because it helped me figure out if I
was actually loading the plugin or not, and might be of help to
someone else who's just starting out with plugins.